+++ comp/logs/agent/config/config_keys.go
@@ -117,8 +117,8 @@ func (l *LogsConfigKeys) devModeUseProto() bool {
 	return l.getConfig().GetBool(l.getConfigKey("dev_mode_use_proto"))
 }
 
-func (l *LogsConfigKeys) httpConnectivityRetryIntervalMax() int {
-	return l.getConfig().GetInt(l.getConfigKey("http_connectivity_retry_interval_max"))
+func (l *LogsConfigKeys) httpConnectivityRetryIntervalMax() float64 {
+	return l.getConfig().GetFloat64(l.getConfigKey("http_connectivity_retry_interval_max"))
 }
 
 func (l *LogsConfigKeys) compressionKind() string {
